delete modal as its own component
